fixed urls

+++ src/main/java/finalproject/soundcloud/model/daos/SongDao.java
@@ -4,11 +4,13 @@ import finalproject.soundcloud.model.dtos.SongDto;
 import finalproject.soundcloud.model.pojos.Song;
 import finalproject.soundcloud.model.pojos.User;
 import finalproject.soundcloud.model.repostitories.SongRepository;
+import finalproject.soundcloud.util.AmazonClient;
 import finalproject.soundcloud.util.exceptions.DoesNotExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.sound.sampled.*;
 import java.io.File;
@@ -30,15 +32,12 @@ public class SongDao {
     @Autowired
     PlaylistDao playlistDao;
     public static final String SONGS_DIR = "D:\\ITtalents\\FinalProject\\";
+    @Transactional
     public ResponseDto rateSong(long songId, User user, boolean clickedLike) throws DoesNotExistException {
-        String sql = "SELECT * FROM songs WHERE song_id LIKE ?";
-         List<Song> songs =  jdbcTemplate.query(sql,new Object[]{songId},new BeanPropertyRowMapper<>(Song.class));
-         if(songs.isEmpty()){ // check if the song exists
-             throw new DoesNotExistException("song");
-         }
+        String sql;
          // check if the song has been liked/disliked before
-         sql = "DELETE FROM users_disliked_songs WHERE song_id = ?";
-         int changed = jdbcTemplate.update(sql,songId);
+         sql = "DELETE FROM users_disliked_songs WHERE song_id = ? AND user_id = ?";
+         int changed = jdbcTemplate.update(sql,songId,user.getId());
          if(changed != 0){
              Song song = songRepository.findById(songId);
              song.setDislikes(song.getDislikes() - 1);
@@ -46,8 +45,8 @@ public class SongDao {
              if(!clickedLike)
                 return new ResponseDto("song removed from disliked!");
          }
-         sql = " DELETE FROM users_liked_songs WHERE song_id = ?";
-         changed = jdbcTemplate.update(sql,songId);
+         sql = " DELETE FROM users_liked_songs WHERE song_id = ? AND user_id = ?";
+         changed = jdbcTemplate.update(sql,songId,user.getId());
          if(changed != 0 ){
              Song song = songRepository.findById(songId);
              song.setLikes(song.getLikes() - 1);
@@ -66,11 +65,13 @@ public class SongDao {
          responseDto.setResponse("action complete!");
          return responseDto;
     }
+
     public ResponseDto repostSong(User user, long songId){
         String sql = "INSERT INTO users_reposts(user_id,song_id) VALUES(?,?)";
         jdbcTemplate.update(sql,user.getId(),songId);
         return new ResponseDto("song was reposted to your profile");
     }
+
     public ResponseDto unpostSong(User user,long songId){
         String sql = "DELETE FROM users_reposts WHERE user_id = ? AND song_id = ?";
         jdbcTemplate.update(sql,user.getId(),songId);
@@ -82,6 +83,7 @@ public class SongDao {
          List<Song> songs= jdbcTemplate.query(sql,new Object[]{user.getId(),songId}, new BeanPropertyRowMapper<>(Song.class));
          return !songs.isEmpty();
     }
+
     public boolean canUploadSong(User user, File song){
         if(user.getUserType() > 2)
             return true;
@@ -129,11 +131,13 @@ public class SongDao {
         return Duration.ofSeconds(Math.round(durationInSeconds));
     }
 
-    public void uploadSong(String name, User user, SongDto dto,File song) {
+    public void uploadSong(String name, User user, SongDto dto,File song) throws IOException {
         String sql = "INSERT INTO songs(user_id,song_name,is_public,file_path,length) " +
                 "VALUES(?,?,?,?,?);";
         jdbcTemplate.update(sql,user.getId(),dto.getSongName(),dto.isPublic(),name,getSongDuration(song).getSeconds());
         //TODO ADD THE SONG IN AWS
+        AmazonClient amazonClient = new AmazonClient();
+        amazonClient.uploadFile(song);
     }
 
     public boolean deleteSong(long songId)  {
